+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Support for redefining function overloads.
 - Add an ocean sample to the `omni.warp` extension.
 - Support for a new `wp.static(expr)` function that allows arbitrary Python expressions to be evaluated at the time of function/kernel definition ([docs](https://nvidia.github.io/warp/codegen.html#static-expressions)).
+- `warp.sim.VBDIntegrator` now supports body-particle collision.
 
 ### Changed